# Optimize various initialization paths