--- a/Gemfile
+++ b/Gemfile
@@ -10,10 +10,6 @@ gem "rake", "~> 13.0", require: false
 # Testing framework
 gem "rspec", "~> 3.0", require: false
 
-# Benchmarking performance
-gem "benchmark", require: false
-gem "benchmark-ips", require: false
-
 group :linting do
   # Linting
   gem "rubocop", "~> 1.72", require: false
@@ -48,4 +44,9 @@ group :development do
 
   # Version changes
   gem "bump", require: false
+
+  # Benchmarking performance
+  gem "benchmark", require: false
+  gem "benchmark-ips", require: false
+  gem "stackprof", require: false
 end