--- a/bin/benchmark
+++ b/bin/benchmark
@@ -43,12 +43,23 @@ end
 `stackprof tmp/stackprof.dump  --d3-flamegraph > tmp/flamegraph.html`
 
 # Baseline:
-#                 real    120.248k (± 6.7%) i/s    (8.32 μs/i) -    600.984k in   5.026862s
-#            real-real    123.672k (± 6.2%) i/s    (8.09 μs/i) -    616.224k in   5.006814s
-#              complex    118.893k (± 4.8%) i/s    (8.41 μs/i) -    597.450k in   5.039148s
-#        string/symbol    273.629k (± 6.4%) i/s    (3.65 μs/i) -      1.373M in   5.042627s
-#       from vector []    289.128k (± 6.4%) i/s    (3.46 μs/i) -      1.463M in   5.085560s
-#      from vector new    379.761k (± 6.6%) i/s    (2.63 μs/i) -      1.891M in   5.010302s
-#                   -@    235.569k (± 6.8%) i/s    (4.25 μs/i) -      1.192M in   5.088220s
-#                  * 3    248.626k (± 6.2%) i/s    (4.02 μs/i) -      1.260M in   5.097549s
-#          various (+)    165.008k (± 4.9%) i/s    (6.06 μs/i) -    831.630k in   5.054346s
+#             real    120.248k (± 6.7%) i/s    (8.32 μs/i) -    600.984k in   5.026862s
+#        real-real    123.672k (± 6.2%) i/s    (8.09 μs/i) -    616.224k in   5.006814s
+#          complex    118.893k (± 4.8%) i/s    (8.41 μs/i) -    597.450k in   5.039148s
+#    string/symbol    273.629k (± 6.4%) i/s    (3.65 μs/i) -      1.373M in   5.042627s
+#   from vector []    289.128k (± 6.4%) i/s    (3.46 μs/i) -      1.463M in   5.085560s
+#  from vector new    379.761k (± 6.6%) i/s    (2.63 μs/i) -      1.891M in   5.010302s
+#               -@    235.569k (± 6.8%) i/s    (4.25 μs/i) -      1.192M in   5.088220s
+#              * 3    248.626k (± 6.2%) i/s    (4.02 μs/i) -      1.260M in   5.097549s
+#      various (+)    165.008k (± 4.9%) i/s    (6.06 μs/i) -    831.630k in   5.054346s
+
+# After optimization of init paths (#13):
+#             real    289.522k (± 3.4%) i/s    (3.45 μs/i) -      1.453M in   5.024349s
+#        real-real    329.834k (± 2.6%) i/s    (3.03 μs/i) -      1.666M in   5.054816s
+#          complex    268.877k (± 4.9%) i/s    (3.72 μs/i) -      1.358M in   5.067001s
+#    string/symbol    383.751k (± 3.3%) i/s    (2.61 μs/i) -      1.932M in   5.039982s
+#   from vector []    356.168k (± 2.6%) i/s    (2.81 μs/i) -      1.791M in   5.032325s
+#  from vector new    897.485k (± 4.3%) i/s    (1.11 μs/i) -      4.543M in   5.072670s
+#               -@    443.206k (± 5.1%) i/s    (2.26 μs/i) -      2.251M in   5.095046s
+#              * 3    487.009k (± 5.2%) i/s    (2.05 μs/i) -      2.464M in   5.077360s
+#      various (+)    234.189k (± 2.3%) i/s    (4.27 μs/i) -      1.172M in   5.008158s