--- a/lib/vector_number.rb
+++ b/lib/vector_number.rb
@@ -176,7 +176,6 @@ class VectorNumber
     @data = values.to_h and return if values.is_a?(VectorNumber)
 
     @data = Hash.new(0)
-
     case values
     when Array
       values.each { |value| add_value_to_data(value) }
@@ -251,32 +250,27 @@ class VectorNumber
   # @since 0.1.0
   def save_options(options, values:)
     @options =
-      case [options, values]
-      in [{} | nil, VectorNumber]
-        values.options
-      in [{} | nil, [*, VectorNumber => vector, *]]
-        vector.options
-      in Hash, VectorNumber
+      case values
+      in VectorNumber
         merge_options(values.options, options)
-      in Hash, [*, VectorNumber => vector, *]
+      in [*, VectorNumber => vector, *]
         merge_options(vector.options, options)
-      in Hash, _ unless options.empty?
-        merge_options(default_options, options)
       else
-        default_options
+        merge_options(DEFAULT_OPTIONS, options)
       end
   end
 
   # @param base_options [Hash{Symbol => Object}]
-  # @param added_options [Hash{Symbol => Object}]
+  # @param added_options [Hash{Symbol => Object}, nil]
   # @return [Hash{Symbol => Object}]
   #
   # @since 0.3.0
   def merge_options(base_options, added_options)
+    return base_options if !added_options || added_options.empty?
     # Optimization for the common case of passing options through #new.
     return base_options if added_options.equal?(base_options)
 
-    base_options.merge(added_options).slice(*known_options)
+    base_options.merge(added_options).slice(*KNOWN_OPTIONS)
   end
 
   # Compact coefficients, calculate size and freeze data.
@@ -287,15 +281,5 @@ class VectorNumber
     @data.delete_if { |_u, c| c.zero? }
     @data.freeze
     @size = @data.size
-  end
-
-  # @since 0.2.0
-  def default_options
-    DEFAULT_OPTIONS
-  end
-
-  # @since 0.2.0
-  def known_options
-    KNOWN_OPTIONS
   end
 end