--- a/spec/vector_number/enumerating_spec.rb
+++ b/spec/vector_number/enumerating_spec.rb
@@ -43,9 +43,9 @@ RSpec.describe VectorNumber::Enumerating do
       end
 
       include_examples "yields values", for_number: :zero_number, values: []
-      include_examples "yields values", for_number: :real_number, values: [[0.i, 1.5r]]
+      include_examples "yields values", for_number: :real_number, values: [[1, 1.5r]]
       include_examples "yields values", for_number: :composite_number,
-                                        values: [["y", 1], [:a, 1], [0.i, 5]]
+                                        values: [["y", 1], [:a, 1], [1, 5]]
     end
   end
 
@@ -65,8 +65,8 @@ RSpec.describe VectorNumber::Enumerating do
     end
 
     include_examples "returns units", for_number: :zero_number, values: []
-    include_examples "returns units", for_number: :real_number, values: [0.i]
-    include_examples "returns units", for_number: :composite_number, values: [0.i, "y", :a]
+    include_examples "returns units", for_number: :real_number, values: [1]
+    include_examples "returns units", for_number: :composite_number, values: [1, "y", :a]
   end
 
   include_examples "has an alias", :keys, :units
@@ -100,7 +100,7 @@ RSpec.describe VectorNumber::Enumerating do
     context "without a block" do
       it "returns plain vector representation without calling #each" do
         # Can't actually test for not calling #each, as objects are frozen and can't be mocked.
-        expect(hash).to eq(0.i => 5, "y" => 1, :a => 1)
+        expect(hash).to eq(1 => 5, "y" => 1, :a => 1)
       end
     end
 
@@ -108,7 +108,7 @@ RSpec.describe VectorNumber::Enumerating do
       let(:block) { ->(u, v) { [u, u.is_a?(Numeric) ? v : v * 0.5] } }
 
       it "returns transformed plain vector representation" do
-        expect(hash).to eq(0.i => 5, "y" => 0.5, :a => 0.5)
+        expect(hash).to eq(1 => 5, "y" => 0.5, :a => 0.5)
       end
     end
   end