--- a/spec/vector_number/new_spec.rb
+++ b/spec/vector_number/new_spec.rb
@@ -162,4 +162,12 @@ RSpec.describe VectorNumber, ".new", :aggregate_failures do
       end
     end
   end
+
+  context "when initializing with unsupported type" do
+    let(:value) { Set.new(1, 2, 3) }
+
+    it "raises ArgumentError" do
+      expect { new_number }.to raise_error ArgumentError
+    end
+  end
 end